Move ponca types to dedicated header file
